add the delete comment function in usePostComment custom hook and apply it on post Details page
